requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
mysql-connector==2.2.9
packaging==24.1
Werkzeug==3.0.3
